chore(auth): removes legacy components/routes
