--- src/Apps/Authentication/Legacy/Server/setReferer.ts
@@ -1,4 +1,0 @@
-export const setReferer = ({ req, res }) => {
-  res.locals.sd.AUTHENTICATION_REFERER = req.header("Referer") || req.hostname
-  return { req, res }
-}
